Create usePersonalBlogpostInfo hook
