+++ packages/lesswrong/components/tagging/TagFilterSettings.tsx
@@ -4,9 +4,9 @@ import { FilterSettings, FilterMode, isCustomFilterMode } from '../../lib/filter
 import { useCurrentUser } from '../common/withUser';
 import { tagStyle } from './FooterTag';
 import { filteringStyles } from './FilterMode';
+import { usePersonalBlogpostInfo } from './usePersonalBlogpostInfo';
 import Card from '@material-ui/core/Card';
 import { userHasNewTagSubscriptions } from '../../lib/betas';
-import { ForumOptions, forumSelect } from '../../lib/forumTypeUtils';
 import { taggingNameCapitalSetting } from '../../lib/instanceSettings';
 
 const styles = (theme: ThemeType): JssStyles => ({
@@ -42,42 +42,6 @@ const styles = (theme: ThemeType): JssStyles => ({
   }
 });
 
-const lwafPersonalBlogpostInfo = {
-  name: "Personal Blog",
-  tooltip: <div>
-    <p><b>Personal Blogposts</b> are posts that don't fit LessWrong's Frontpage Guidelines. They get less visibility by default. The frontpage guidelines are:</p>
-    <ul>
-      <li><em>Timelessness</em>. Will people still care about this in 5 years?</li>
-      <li><em>Avoid political topics</em>. They're important to discuss sometimes, but we try to avoid it on LessWrong.</li>
-      <li><em>General Appeal</em>. Is this a niche post that only a small fraction of users will care about?</li>
-    </ul>
-  </div>
-}
-
-const personalBlogpostInfo: ForumOptions<{name: string, tooltip: JSX.Element}> = {
-  LessWrong: lwafPersonalBlogpostInfo,
-  AlignmentForum: lwafPersonalBlogpostInfo,
-  EAForum: {
-    name: 'Personal',
-    tooltip: <div>
-      <div>
-        By default, the home page only displays Frontpage Posts, which are selected by moderators as especially interesting or useful to people with interest in doing good effectively. Personal posts get to have looser standards of relevance, and may include topics that could lead to more emotive or heated discussion (e.g. politics), which are generally excluded from Frontpage.
-      </div>
-    </div>
-  },
-  default: {
-    name: 'Personal',
-    tooltip: <div>
-      <div>
-        By default, the home page only displays Frontpage Posts, which are selected by moderators as especially interesting or useful to people with interest in doing good effectively. Personal posts get to have looser standards of relevance, and may include topics that could lead to more emotive or heated discussion (e.g. politics), which are generally excluded from Frontpage.
-      </div>
-    </div>
-  },
-}
-
-const personalBlogpostName = forumSelect(personalBlogpostInfo).name
-const personalBlogpostTooltip = forumSelect(personalBlogpostInfo).tooltip
-
 /**
  * Adjust the weighting the frontpage gives to posts with the given tags
  *
@@ -99,6 +63,11 @@ const TagFilterSettings = ({
 }) => {
   const { AddTagButton, FilterMode, Loading, LWTooltip, ContentStyles } = Components
   const currentUser = useCurrentUser()
+
+  const {
+    name: personalBlogpostName,
+    tooltip: personalBlogpostTooltip,
+  } = usePersonalBlogpostInfo();
 
   const personalBlogpostCard = <Card>
     <ContentStyles contentType="comment" className={classes.personalTooltip}>
